рефакторинг структуры
