+++ lib/api/constants/places_api_urls.dart
@@ -1,0 +1,8 @@
+/// Адреса для работы с местами
+class PlacesApiUrls {
+  static const String places = '/place';
+  static const String filteredPlaces = '/filtered_places';
+  static const String uploadFile = '/upload_file';
+
+  PlacesApiUrls._();
+}
