+++ lib/features/app/di/app_scope.dart
@@ -3,12 +3,12 @@ import 'dart:ui';
 import 'package:dio/adapter.dart';
 import 'package:dio/dio.dart';
 import 'package:elementary/elementary.dart';
+import 'package:places_elementary/api/service/places_api.dart';
 import 'package:places_elementary/config/app_config.dart';
 import 'package:places_elementary/config/environment/environment.dart';
-import 'package:places_elementary/features/feature_places/domain/repository/api/service/place_api.dart';
-import 'package:places_elementary/features/feature_places/domain/repository/place_repository.dart';
-import 'package:places_elementary/features/feature_places/service/places_service.dart';
 import 'package:places_elementary/features/navigation/service/coordinator.dart';
+import 'package:places_elementary/features/places/domain/repository/places_repository.dart';
+import 'package:places_elementary/features/places/service/places_service.dart';
 import 'package:places_elementary/util/default_error_handler.dart';
 
 /// Scope of dependencies which need through all app's life.
@@ -19,7 +19,7 @@ class AppScope implements IAppScope {
   late final Coordinator _coordinator;
 
   // Места
-  late final PlaceApi _placeApi;
+  late final PlacesApi _placesApi;
   late final PlacesRepository _placesRepository;
   late final PlacesService _placesService;
 
@@ -89,8 +89,8 @@ class AppScope implements IAppScope {
 
   /// Работа с местами
   PlacesService _initPlacesService(Dio dio) {
-    _placeApi = PlaceApi(dio);
-    _placesRepository = PlacesRepository(_placeApi);
+    _placesApi = PlacesApi(dio);
+    _placesRepository = PlacesRepository(_placesApi);
 
     return PlacesService(_placesRepository);
   }
