--- lib/features/feature_places/domain/repository/api/constants/place_api_urls.dart
@@ -1,8 +1,0 @@
-/// Адреса для работы с местами
-class PlaceApiUrls {
-  static const String places = '/place';
-  static const String filteredPlaces = '/filtered_places';
-  static const String uploadFile = '/upload_file';
-
-  PlaceApiUrls._();
-}
