+++ lib/features/navigation/domain/entity/app_coordinate.dart
@@ -1,10 +1,10 @@
 import 'package:places_elementary/features/debug/screens/debug_screen/debug_screen.dart';
-import 'package:places_elementary/features/feature_favorites/screens/favorites_screen.dart';
-import 'package:places_elementary/features/feature_main/screens/main_screen/main_screen.dart';
-import 'package:places_elementary/features/feature_map/screens/map_screen.dart';
-import 'package:places_elementary/features/feature_places/screens/places_screen/places_screen.dart';
-import 'package:places_elementary/features/feature_settings/screens/settings_screen.dart';
+import 'package:places_elementary/features/favorites/screens/favorites_screen.dart';
+import 'package:places_elementary/features/main/screens/main_screen/main_screen.dart';
+import 'package:places_elementary/features/map/screens/map_screen.dart';
 import 'package:places_elementary/features/navigation/domain/entity/coordinate.dart';
+import 'package:places_elementary/features/places/screens/places_screen/places_screen.dart';
+import 'package:places_elementary/features/settings/screens/settings_screen.dart';
 import 'package:places_elementary/features/temp/screens/temp_screen/temp_screen.dart';
 
 /// A set of routes for the entire app.
