+++ lib/features/places/domain/entity/card_type.dart
