+++ lib/features/places/domain/entity/object_position.dart
