+++ lib/features/places/domain/entity/place.freezed.dart
