+++ lib/features/places/domain/entity/place_type.dart
