+++ lib/features/places/domain/entity/search_filter.dart
