+++ lib/features/places/screens/places_screen/places_screen_model.dart
@@ -1,0 +1,13 @@
+import 'package:elementary/elementary.dart';
+import 'package:places_elementary/features/places/domain/entity/place.dart';
+import 'package:places_elementary/features/places/service/places_service.dart';
+
+/// Модель для PlacesScreen
+class PlacesScreenModel extends ElementaryModel {
+  final PlacesService placesService;
+
+  PlacesScreenModel(this.placesService);
+
+  /// Получить все места
+  Future<List<Place>> getAllPlaces() => placesService.getAllPlaces();
+}
