+++ lib/features/places/widgets/card_content_type.dart
